server: Integrate reconnect tasks into `ChiaServer`
